// src/services/moguding-backend/models/PlatformInfoUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type PlatformInfoUpdateRequest = {
    clockType?: string;
    coin?: number;
    id?: number;
    name?: string;
    status?: number;
    timeType?: string;
    url?: string;
};